backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pillow==10.4.0
numpy<2
onnxruntime==1.17.3
rembg==2.0.68
python-multipart==0.0.9
